chore: move variant generator func to Palette
