+++ src/lib/palette.svelte.ts
@@ -1,8 +1,11 @@
 import { colord, extend, random, type Colord } from 'colord';
 import namesPlugin from 'colord/plugins/names';
 import { getContext, setContext } from 'svelte';
+import { linear, quadIn, quadInOut, quadOut } from 'svelte/easing';
 
 extend([namesPlugin]);
+
+const easingFns = { linear, quadInOut, quadIn, quadOut };
 
 export interface Color {
 	source: Swatch;
@@ -23,6 +26,30 @@ export interface Swatch {
 
 export class Palette {
 	colors = $state<Color[]>([]);
+
+	newColor() {
+		this.colors.push(this.stringToColor(random().toHex()));
+	}
+
+	stringToColor(colorString: string): Color {
+		return this.colordToColor(colord(colorString));
+	}
+
+	getClosestCSSColorName(colorString: string): string {
+		return colord(colorString).toName({ closest: true }) ?? '';
+	}
+
+	isColorValid(colorString: string): boolean {
+		return colord(colorString).isValid();
+	}
+
+	hslToColor(hsl: { h: number; s: number; l: number }): Color {
+		return this.colordToColor(colord({ ...hsl }));
+	}
+
+	updateColorVariants(color: Color) {
+		color.variants = this.generateVariants(color);
+	}
 
 	private colordToSwatch(colord: Colord): Swatch {
 		return {
@@ -45,24 +72,29 @@ export class Palette {
 		};
 	}
 
-	newColor() {
-		this.colors.push(this.stringToColor(random().toHex()));
+	private generateVariants(color: Color): Swatch[] {
+		const colorVariants: Swatch[] = [];
+		const hsl = color.source.hsl;
+
+		const MIN_LIGHTNESS = 0;
+		const MAX_LIGHTNESS = 100;
+		const RANGE = MAX_LIGHTNESS - MIN_LIGHTNESS;
+
+		for (let i = 0; i < color.steps; i++) {
+			const progress = i / (color.steps - 1);
+			const easedProgress = this.getEasingValue(color.easingFn, progress);
+			const rangePercentage = RANGE * (1 - easedProgress);
+			const lightness = MIN_LIGHTNESS + rangePercentage;
+
+			const variant = this.hslToColor({ ...hsl, l: lightness });
+			colorVariants.push(variant.source);
+		}
+
+		return colorVariants;
 	}
 
-	stringToColor(colorString: string): Color {
-		return this.colordToColor(colord(colorString));
-	}
-
-	getClosestCSSColorName(colorString: string): string {
-		return colord(colorString).toName({ closest: true }) ?? '';
-	}
-
-	isColorValid(colorString: string): boolean {
-		return colord(colorString).isValid();
-	}
-
-	hslToColor(hsl: { h: number; s: number; l: number }): Color {
-		return this.colordToColor(colord({ ...hsl }));
+	private getEasingValue(easingFn: string, progress: number): number {
+		return easingFns[easingFn as keyof typeof easingFns](progress);
 	}
 }
